W-12724722: Extract tck from mule-api (test classes used in extensions-api and mule modules)
